electron-demo initial commit
