--- electron-node-datachannel-demo/src/preload.js
@@ -1,2 +1,0 @@
-// See the Electron documentation for details on how to use preload scripts:
-// https://www.electronjs.org/docs/latest/tutorial/process-model#preload-scripts
