+++ examples/electron-demo/webpack.renderer.config.js
@@ -1,0 +1,13 @@
+const rules = require('./webpack.rules');
+
+rules.push({
+    test: /\.css$/,
+    use: [{ loader: 'style-loader' }, { loader: 'css-loader' }],
+});
+
+module.exports = {
+    // Put your normal webpack config below here
+    module: {
+        rules,
+    },
+};
